Moved getCloudProviderFromWorkspace to libs/workspace-utils.ts.
