+++ src/components/workspace-utils.js
@@ -13,7 +13,6 @@ import TooltipTrigger from 'src/components/TooltipTrigger'
 import { ReactComponent as CloudAzureLogo } from 'src/images/cloud_azure_icon.svg'
 import { ReactComponent as CloudGcpLogo } from 'src/images/cloud_google_icon.svg'
 import { Ajax } from 'src/libs/ajax'
-import { cloudProviderTypes } from 'src/libs/ajax/ajax-common'
 import colors from 'src/libs/colors'
 import { getConfig } from 'src/libs/config'
 import { reportError, withErrorReporting } from 'src/libs/error'
@@ -504,9 +503,4 @@ export const NoWorkspacesMessage = ({ onClick }) => {
       }, ['What\'s a workspace?'])
     ])
   ])
-}
-
-//TODO: Workspace should contain it's cloud provider in a better fashion.
-export const getCloudProviderFromWorkspace = workspace => {
-  return workspace.azureContext ? cloudProviderTypes.AZURE : cloudProviderTypes.GCP
 }
